+++ python/micromode/constants.py
@@ -1,0 +1,4 @@
+"""Physical constants used by MicroMode."""
+
+C_0 = 2.997_924_58e14
+EPSILON_0 = 8.854_187_812_800_384e-18